+++ src/app/app-routing.module.ts
@@ -29,6 +29,7 @@ import { CanDeactivateDashboard } from './guards/can-deactivate-dashboard.guard'
 import { WarningsModuleComponent } from './dashboard/warnings-module/warnings-module.component';
 import { TimersModuleComponent } from './dashboard/timers-module/timers-module.component';
 import { BadgesComponent } from './dashboard/badges/badges.component';
+import { CratesComponent } from './crates/crates.component';
 
 const routes: Routes = [
   { path: '', component: HomeComponent },
@@ -43,6 +44,7 @@ const routes: Routes = [
 
   { path: 'dashboard', component: DashboardComponent, canActivate: [DashboardAuthGuard] },
   { path: 'dashboard/badges', component: BadgesComponent, canActivate: [DashboardAuthGuard] },
+  { path: 'dashboard/crates', component: CratesComponent, canActivate: [DashboardAuthGuard] },
   { path: 'dashboard/xp-card', component: XPCardComponent, canActivate: [DashboardAuthGuard] },
 
   { path: 'servers/:id', component: GuildComponent, canActivate: [GuildAuthGuard] },