+++ src/environments/environment.prod.ts
@@ -2,7 +2,7 @@ export const environment = {
   production: true,
   endpoint: '/api',
   url: 'https://3pg.xyz',
-  discordURL: 'https://discordapp.com/invite/xRT6Fz5',
+  discordURL: 'https://discord.com/invite/xRT6Fz5',
   stripePublicKey: 'pk_live_kyJ00K40kgFhHMYzO7Bc9YjV00sVDW6IlV',
   version: 'v2.0.1a',
   voteURLs: [